+++ conanfile.py
@@ -60,7 +60,7 @@ class BayesNetConan(ConanFile):
         self.requires("libtorch/2.7.1")
         self.requires("nlohmann_json/3.11.3")
         self.requires("folding/1.1.2")  # Custom package
-        self.requires("fimdlp/2.1.1")  # Custom package
+        self.requires("fimdlp/2.1.2")  # Custom package
 
     def build_requirements(self):
         self.build_requires("cmake/[>=3.27]")